+++ packages/nimble-components/src/table-column/text/tests/table-column-text.spec.ts
@@ -42,6 +42,16 @@ describe('TableColumnText', () => {
         await disconnect();
     });
 
+    it('should export its tag', () => {
+        expect(tableColumnTextTag).toBe('nimble-table-column-text');
+    });
+
+    it('can construct an element instance', () => {
+        expect(
+            document.createElement('nimble-table-column-text')
+        ).toBeInstanceOf(TableColumnText);
+    });
+
     it('reports column configuration valid', async () => {
         await connect();
         await waitForUpdatesAsync();